move sas to metrics
